Incremental loads from flat files don't work properly, remove that functionality

+++ moves3.py
@@ -8,7 +8,8 @@ bucket = s3.Bucket(SOURCE_BUCKET)
 objects = bucket.objects.all()
 
 for obj in objects:
-    print("Moving " + obj.key)
-    source_path = SOURCE_BUCKET + "/" + obj.key
-    s3.Object(DEST_BUCKET, obj.key).copy_from(CopySource=source_path)
-    s3.Object(SOURCE_BUCKET, obj.key).delete()
+    if obj.key == '201702.csv' or obj.key == '201703.csv' or obj.key == '201704.csv':
+        print("Moving " + obj.key)
+        source_path = SOURCE_BUCKET + "/" + obj.key
+        s3.Object(DEST_BUCKET, obj.key).copy_from(CopySource=source_path)
+        s3.Object(SOURCE_BUCKET, obj.key).delete()
